Resolve #10 and #9
this will require a save-for-previous
numerous small changes:
-AMC dispatcher receiver now encodes source address as well as port
(originally just appended port) in case its ever missing from the send
(unlikely). The original removal seems to have been to fix #5, this
change just restores the functionality to a safer state
-network send now does a series of checks for udp port. If included in
the string and a force boolean (not exposed to user right now) is not
set, it will use the port in the string. If no port in the string it
goes back to the old behavior (if non-0, use, if 0 use global default
54321).
-Created a new FGV for network send which compares the target address to
all local IPs and tries to get the closest network match. It also caches
this result for speed.
-Tweaked multi-target example to test this